requirements.txt
flask
psycopg2
python-dotenv
cloudinary
cloudinary-storage
